turned over Gather inheritance hierarchy
