+++ inference-engine/src/transformations/include/transformations/op_conversions/convert_gather_downgrade.hpp
@@ -12,6 +12,7 @@ namespace ngraph {
 namespace pass {
 
 class TRANSFORMATIONS_API ConvertGather7ToGather1;
+class TRANSFORMATIONS_API ConvertGather8ToGather7;
 
 }  // namespace pass
 }  // namespace ngraph
@@ -24,4 +25,14 @@ class ngraph::pass::ConvertGather7ToGather1 : public ngraph::pass::MatcherPass {
 public:
     NGRAPH_RTTI_DECLARATION;
     ConvertGather7ToGather1();
+};
+
+/**
+ * @ingroup ie_transformation_common_api
+ * @brief ConvertGather8ToGather7 converts v8::Gather into v7::Gather.
+ */
+class ngraph::pass::ConvertGather8ToGather7 : public ngraph::pass::MatcherPass {
+public:
+    NGRAPH_RTTI_DECLARATION;
+    ConvertGather8ToGather7();
 };
