+++ inference-engine/src/transformations/include/transformations/op_conversions/convert_gather_upgrade.hpp
@@ -13,6 +13,8 @@ namespace pass {
 
 class TRANSFORMATIONS_API ConvertGather1ToGather7;
 
+class TRANSFORMATIONS_API ConvertGather7ToGather8;
+
 }  // namespace pass
 }  // namespace ngraph
 
@@ -24,4 +26,14 @@ class ngraph::pass::ConvertGather1ToGather7 : public ngraph::pass::MatcherPass {
 public:
     NGRAPH_RTTI_DECLARATION;
     ConvertGather1ToGather7();
+};
+
+/**
+ * @ingroup ie_transformation_common_api
+ * @brief ConvertGather7ToGather8 converts v7::Gather into v8::Gather.
+ */
+class ngraph::pass::ConvertGather7ToGather8 : public ngraph::pass::MatcherPass {
+public:
+    NGRAPH_RTTI_DECLARATION;
+    ConvertGather7ToGather8();
 };
