+++ inference-engine/src/transformations/src/transformations/common_optimizations/eliminate_unsqueeze_gather.cpp
@@ -19,7 +19,7 @@ ngraph::pass::EliminateUnsqueezeGather::EliminateUnsqueezeGather() {
     const auto unsqueeze = ngraph::pattern::wrap_type<ngraph::opset6::Unsqueeze>({unsqueezeInput, unsqueezeAxis}, pattern::consumers_count(1));
     const auto gatherIndices = ngraph::opset6::Constant::create(ngraph::element::i64, ngraph::Shape{}, {0});
     const auto gatherAxis = ngraph::pattern::any_input();
-    const auto gather = ngraph::pattern::wrap_type<ngraph::op::util::GatherBase>({unsqueeze, gatherIndices, gatherAxis});
+    const auto gather = ngraph::pattern::wrap_type<ngraph::op::v8::Gather>({unsqueeze, gatherIndices, gatherAxis});
 
     ngraph::matcher_pass_callback callback = [=](ngraph::pattern::Matcher& m) {
         auto& patternValue = m.get_pattern_value_map();
