+++ inference-engine/src/transformations/src/transformations/op_conversions/convert_gather_downgrade.cpp
@@ -2,15 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "transformations/op_conversions/convert_gather_v7_to_gather_v1.hpp"
+#include "transformations/op_conversions/convert_gather_downgrade.hpp"
 #include <ngraph/opsets/opset1.hpp>
 #include <ngraph/opsets/opset7.hpp>
+#include <ngraph/opsets/opset8.hpp>
 #include <ngraph/rt_info.hpp>
 #include <ngraph/pattern/op/wrap_type.hpp>
 
 #include "itt.hpp"
 
 NGRAPH_RTTI_DEFINITION(ngraph::pass::ConvertGather7ToGather1, "ConvertGather7ToGather1", 0);
+NGRAPH_RTTI_DEFINITION(ngraph::pass::ConvertGather8ToGather7, "ConvertGather8ToGather7", 0);
 
 ngraph::pass::ConvertGather7ToGather1::ConvertGather7ToGather1() {
     MATCHER_SCOPE(ConvertGather7ToGather1);
@@ -37,5 +39,31 @@ ngraph::pass::ConvertGather7ToGather1::ConvertGather7ToGather1() {
     };
 
     auto m = std::make_shared<pattern::Matcher>(gather_v7, matcher_name);
+    register_matcher(m, callback);
+}
+
+ngraph::pass::ConvertGather8ToGather7::ConvertGather8ToGather7() {
+    MATCHER_SCOPE(ConvertGather8ToGather7);
+
+    auto gather_v8 = pattern::wrap_type<ngraph::opset8::Gather>();
+
+    ngraph::matcher_pass_callback callback = [=](pattern::Matcher& m) {
+        auto gather_v8_node = std::dynamic_pointer_cast<ngraph::opset7::Gather>(m.get_match_root());
+        if (!gather_v8_node)
+            return false;
+
+        auto data_input = gather_v8_node->input_value(0);
+        auto indices_input = gather_v8_node->input_value(1);
+        auto axis_input = gather_v8_node->input_value(2);
+        auto batch_dims = gather_v8_node->get_batch_dims();
+
+        auto gather_v7 = std::make_shared<ngraph::opset7::Gather>(data_input, indices_input, axis_input, batch_dims);
+        gather_v7->set_friendly_name(gather_v8_node->get_friendly_name());
+        ngraph::copy_runtime_info(gather_v8_node, gather_v7);
+        ngraph::replace_node(gather_v8_node, gather_v7);
+        return true;
+    };
+
+    auto m = std::make_shared<pattern::Matcher>(gather_v8, matcher_name);
     register_matcher(m, callback);
 }
